Fix-review-comment-Viewport action corners customization
