+++ extensions/default/src/Components/ViewportActionCorners/index.tsx
@@ -1,0 +1,4 @@
+import ViewportActionCorners, { ViewportActionCornersLocations } from './ViewportActionCorners';
+
+export { ViewportActionCornersLocations };
+export default ViewportActionCorners;
