+++ extensions/default/src/customizations/viewportActionCornersCustomization.ts
@@ -1,0 +1,5 @@
+import ViewportActionCorners from '../Components/ViewportActionCorners';
+
+export default {
+  'ui.ViewportActionCorner': ViewportActionCorners,
+};
